refactor[#17] : useDropdown 으로 도메인 로직 추상화
