+++ beta-reader/src/shared/ui/Dropdown/lib/index.ts
@@ -1,0 +1,1 @@
+export * from "./useDropdown";
